scripts/requirements.txt
PyNaCl>=1.5.0
base58>=2.1
requests>=2.31
qrcode>=7.4
Pillow>=10.0
pyzbar>=0.1.9
